Refactor: Extract shared resource registration logic into reusable service
