+++ src/trpc/routers/resources.ts
@@ -1,32 +1,16 @@
 import z from 'zod';
-import z3 from 'zod3';
 
 import { createTRPCRouter, publicProcedure } from '../trpc';
 
-import { scrapeOriginData } from '@/services/scraper';
 import {
   getResourceByAddress,
   listResources,
   searchResources,
   searchResourcesSchema,
-  upsertResource,
 } from '@/services/db/resources';
-import { upsertOrigin } from '@/services/db/origin';
-import { upsertResourceResponse } from '@/services/db/resource-responses';
+import { registerResource } from '@/services/resources/registration';
 
 import { ethereumAddressSchema } from '@/lib/schemas';
-import {
-  EnhancedPaymentRequirementsSchema,
-  parseX402Response,
-} from '@/lib/x402/schema';
-import { formatTokenAmount } from '@/lib/token';
-import { getOriginFromUrl } from '@/lib/url';
-
-import { Methods } from '@/types/x402';
-
-import type { AcceptsNetwork } from '@prisma/client';
-import { x402ResponseSchema } from 'x402/types';
-import { getFaviconUrl } from '@/lib/favicon';
 
 export const resourcesRouter = createTRPCRouter({
   list: {
@@ -65,128 +49,7 @@ export const resourcesRouter = createTRPCRouter({
       })
     )
     .mutation(async ({ input }) => {
-      let parseErrorData: {
-        parseErrors: string[];
-        data: unknown;
-      } | null = null;
-
-      for (const method of [Methods.GET, Methods.POST]) {
-        // ping resource
-        const response = await fetch(input.url, {
-          method,
-          headers: input.headers,
-          body: input.body ? JSON.stringify(input.body) : undefined,
-        });
-
-        // if it doesn't respond with a 402, return error
-        if (response.status !== 402) {
-          continue;
-        }
-
-        const data = (await response.json()) as unknown;
-
-        const baseX402ParsedResponse = x402ResponseSchema
-          .omit({
-            error: true,
-          })
-          .extend({
-            error: z3.string().optional(),
-            accepts: z3.array(EnhancedPaymentRequirementsSchema).optional(),
-          })
-          .safeParse(data);
-        if (!baseX402ParsedResponse.success) {
-          console.error(baseX402ParsedResponse.error.issues);
-          parseErrorData = {
-            parseErrors: baseX402ParsedResponse.error.issues.map(
-              issue => `${issue.path.join('.')}: ${issue.message}`
-            ),
-            data,
-          };
-          continue;
-        }
-
-        const origin = getOriginFromUrl(input.url);
-
-        const {
-          og,
-          metadata,
-          origin: scrapedOrigin,
-        } = await scrapeOriginData(origin);
-
-        await upsertOrigin({
-          origin: origin,
-          title: metadata?.title ?? og?.ogTitle,
-          description: metadata?.description ?? og?.ogDescription,
-          favicon: og?.favicon
-            ? getFaviconUrl(og.favicon, scrapedOrigin)
-            : undefined,
-          ogImages:
-            og?.ogImage?.map(image => ({
-              url: image.url,
-              height: image.height,
-              width: image.width,
-              title: og.ogTitle,
-              description: og.ogDescription,
-            })) ?? [],
-        });
-
-        // upsert the resource
-        const resource = await upsertResource({
-          resource: input.url.toString(),
-          type: 'http',
-          x402Version: baseX402ParsedResponse.data.x402Version,
-          lastUpdated: new Date(),
-          accepts:
-            baseX402ParsedResponse.data.accepts?.map(accept => ({
-              ...accept,
-              network: accept.network.replace('-', '_') as AcceptsNetwork,
-              maxAmountRequired: accept.maxAmountRequired,
-              outputSchema: accept.outputSchema!,
-              extra: accept.extra,
-            })) ?? [],
-        });
-
-        if (!resource) {
-          continue;
-        }
-
-        // parse the response
-        let enhancedParseWarnings: string[] | null = null;
-        const parsedResponse = parseX402Response(data);
-        if (parsedResponse.success) {
-          await upsertResourceResponse(
-            resource.resource.id,
-            parsedResponse.data
-          );
-        } else {
-          enhancedParseWarnings = parsedResponse.errors;
-        }
-
-        return {
-          error: false as const,
-          resource,
-          accepts: {
-            ...resource.accepts,
-            maxAmountRequired: formatTokenAmount(
-              resource.accepts.maxAmountRequired
-            ),
-          },
-          enhancedParseWarnings,
-          response: data,
-        };
-      }
-
-      if (parseErrorData) {
-        return {
-          error: true as const,
-          type: 'parseErrors' as const,
-          parseErrorData,
-        };
-      }
-
-      return {
-        error: true as const,
-        type: 'no402' as const,
-      };
+      // Use shared registration service
+      return await registerResource(input);
     }),
 });
